+++ addons/test_mail/tests/__init__.py
@@ -7,12 +7,11 @@ from . import test_mail_activity
 from . import test_mail_followers
 from . import test_mail_message
 from . import test_mail_mail
-from . import test_mail_cc
 from . import test_mail_channel
 from . import test_mail_gateway
 from . import test_mail_redirect
 from . import test_mail_template
-from . import test_message_composerr
+from . import test_message_composer
 from . import test_message_management
 from . import test_message_post
 from . import test_message_track
